+++ src/ScEventSubscription.ts
@@ -9,15 +9,19 @@ export type ScEventCallbackFunc = (
 
 export enum ScEventType {
   Unknown = "unknown",
-  AddOutgoingEdge = "add_outgoing_edge",
-  AddIngoingEdge = "add_ingoing_edge",
-  RemoveOutgoingEdge = "remove_outgoing_edge",
-  RemoveIngoingEdge = "remove_ingoing_edge",
-  RemoveElement = "delete_element",
-  ChangeContent = "content_change",
+  AfterGenerateConnector = "sc_event_after_generate_connector",
+  AfterGenerateOutgoingArc = "sc_event_after_generate_outgoing_arc",
+  AfterGenerateIncomingArc = "sc_event_after_generate_incoming_arc",
+  AfterGenerateEdge = "sc_event_after_generate_edge",
+  BeforeEraseConnector = "sc_event_before_erase_connector",
+  BeforeEraseOutgoingArc = "sc_event_before_erase_outgoing_arc",
+  BeforeEraseIncomingArc = "sc_event_before_erase_incoming_arc",
+  BeforeEraseEdge = "sc_event_before_erase_edge",
+  BeforeEraseElement = "sc_event_before_erase_element",
+  BeforeChangeLinkContent = "sc_event_before_change_link_content",
 }
 
-export class ScEvent {
+export class ScEventSubscription {
   private _id: number = 0;
   private _type: ScEventType | null = null;
   private _callback: ScEventCallbackFunc | null = null;