+++ docs/changelog.md
@@ -5,6 +5,56 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Breaking changes
+- This version is compatible with version of the sc-machine 0.10.0. All API methods were redesigned. Incorrect ones were removed, new ones were added. See table below, to learn more about changes.
+
+| Deprecated method                      | Substitution method                       | 
+|----------------------------------------|-------------------------------------------|
+| checkElements                          | getElementsTypes                           |
+| createElements                         | generateElements                          |
+| createNode                             | generateNode                              |
+| createLink                             | generateLink                              |
+| createEdge                             | generateConnector                         |
+| createElementsBySCs                    | generateElementsBySCs                     | 
+| deleteElements                         | eraseElements                             |
+| getLinksByContents                     | searchLinksByContents                     |
+| getLinksByContentSubstrings            | searchLinksByContentSubstrings            |
+| getLinksContentsByContentSubstrings    | searchLinkContentsByContentSubstrings     |
+| templateSearch                         | searchByTemplate                          |
+| templateGenerate                       | generateByTemplate                        |
+| tripleWithRelation                     | quintuple                                 |
+| eventsCreate                           | createElementaryEventSubscriptions        |
+| eventsDestroy                          | destroyElementaryEventSubscriptions       |
+| findKeynodes                           | searchKeynodes                            |
+
+- `ScEvent` class was renamed to `ScEventSubscription`, `ScEventParams` class was renamed to `ScEventSubscriptionParams`, and sc-event types in `ScEventType` were changed according the table below.
+
+  | Removed sc-event type  | Substitution sc-event type  |
+  |------------------------|-----------------------------|
+  | AddOutgoingEdge        | AfterGenerateOutgoingArc    |
+  | AddIngoingEdge         | AfterGenerateIncomingArc    |
+  | RemoveOutgoingEdge     | BeforeEraseOutgoingArc      |
+  | RemoveIngoingEdge      | BeforeEraseIncomingArc      |
+  | RemoveElement          | BeforeEraseElement          |
+  | ChangeContent          | BeforeChangeLinkContent     |
+
+- New sc-event types: `AfterGenerateConnector`, `AfterGenerateEdge`, `BeforeEraseConnector`, `BeforeEraseEdge` were added.
+
+### Added
+- sc-event types: `AfterGenerateConnector`, `AfterGenerateOutgoingArc`, `AfterGenerateIncomingArc`, `AfterGenerateEdge`, `BeforeEraseConnector`, `BeforeEraseOutgoingArc`, `BeforeEraseIncomingArc`, `BeforeEraseEdge`, `BeforeEraseElement`, `BeforeChangeLinkContent`
+- ScTemplate methods: `quintuple`
+- ScClient methods: `getElementsTypes`, `generateElements`, `generateNode`, `generateLink`, `generateConnector`, `generateElementsBySCs`, `eraseElements`, `searchLinksByContents`, `searchLinksByContentSubstrings`, `searchLinkContentsByContentSubstrings`, `searchByTemplate`, `generateByTemplate`, `createElementaryEventSubscriptions`, `destroyElementaryEventSubscriptions`, `searchKeynodes`
+
+### Changed
+- `ScEvent` class was renamed to `ScEventSubscription`
+- `ScEventParams` class was renamed to `ScEventSubscriptionParams`
+
+### Deprecated
+- ScTemplate methods: `tripleWithRelation`
+- ScClient methods: `checkElements`, `createElements`, `createNode`, `createLink`, `createEdge`, `createElementsBySCs`, `deleteElements`, `getLinksByContents`, `getLinksByContentSubstrings`, `getLinksContentsByContentSubstrings`, `templateSearch`, `templateGenerate`, `eventsCreate`, `eventsDestroy`, `findKeynodes`
+
+### Removed
+- sc-event types: `AddOutgoingEdge`, `AddIngoingEdge`, `RemoveOutgoingEdge`, `RemoveIngoingEdge`, `RemoveElement`, `ChangeContent`
 
 ## [v0.4.2]
 ### Fixed